requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.7.2
APScheduler==3.10.4
cryptography==41.0.7
Werkzeug==2.3.7
Jinja2==3.1.2
supabase==1.0.4